+++ changelog.md
@@ -1,5 +1,6 @@
 ## Changelog
 
+- There is now a query_filter option for model configs that lets you filter a model's results query before it's constructed
 - Relationship columns now respect WHERE filters in your Eloquent model
 - Bugfix: BelongsToMany filters now work with table prefixes
 - Bugfix: Non-string name_fields and search_fields no longer break select2